requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-groq==0.1.9
unstructured==0.15.0
selenium==4.21.1
chromadb==0.4.0
streamlit==1.25.0
pandas==2.1.1
python-dotenv==1.0.0